requirements.txt
pyserial==3.5
questionary==2.0.0
